+++ apworld/brotato/constants.py
@@ -105,3 +105,4 @@ SHOP_ITEM_LOCATION_TEMPLATE = "{tier} Shop Item {num}"
 # Region name string templates
 CRATE_DROP_GROUP_REGION_TEMPLATE = "Loot Crate Group {num}"
 LEGENDARY_CRATE_DROP_GROUP_REGION_TEMPLATE = "Legendary Loot Crate Group {num}"
+CHARACTER_REGION_TEMPLATE = "In-Game ({char})"